# plan: refine the design of schema