--- a/plan/logical_plan_builder.go
+++ b/plan/logical_plan_builder.go
@@ -432,7 +432,7 @@ func (b *planBuilder) buildSelection(p LogicalPlan, where ast.ExprNode, AggMappe
 				} else {
 					// If there is condition which is always false, return dual plan directly.
 					dual := LogicalTableDual{}.init(b.ctx)
-					dual.SetSchema(p.Schema().Clone())
+					dual.SetSchema(p.Schema())
 					return dual
 				}
 			}
@@ -443,7 +443,6 @@ func (b *planBuilder) buildSelection(p LogicalPlan, where ast.ExprNode, AggMappe
 		return p
 	}
 	selection.Conditions = expressions
-	selection.SetSchema(p.Schema().Clone())
 	selection.SetChildren(p)
 	return selection
 }
@@ -628,9 +627,8 @@ func (b *planBuilder) buildProjection4Union(u *LogicalUnionAll) {
 			proj.SetChildren(child)
 			u.children[childID] = proj
 		}
-		u.children[childID].SetSchema(unionSchema.Clone())
+		u.children[childID].(*LogicalProjection).SetSchema(unionSchema.Clone())
 	}
-	u.SetSchema(unionSchema)
 }
 
 func (b *planBuilder) buildUnion(union *ast.UnionStmt) LogicalPlan {
@@ -695,7 +693,6 @@ func (b *planBuilder) buildSort(p LogicalPlan, byItems []*ast.ByItem, aggMapper 
 	}
 	sort.ByItems = exprs
 	sort.SetChildren(p)
-	sort.SetSchema(p.Schema().Clone())
 	return sort
 }
 
@@ -746,7 +743,6 @@ func (b *planBuilder) buildLimit(src LogicalPlan, limit *ast.Limit) LogicalPlan 
 		Count:  count,
 	}.init(b.ctx)
 	li.SetChildren(src)
-	li.SetSchema(src.Schema().Clone())
 	return li
 }
 
@@ -1549,7 +1545,6 @@ func (b *planBuilder) buildSelect(sel *ast.SelectStmt) LogicalPlan {
 
 func (b *planBuilder) buildTableDual() LogicalPlan {
 	dual := LogicalTableDual{RowCount: 1}.init(b.ctx)
-	dual.SetSchema(expression.NewSchema())
 	return dual
 }
 
@@ -1641,7 +1636,6 @@ func (b *planBuilder) buildDataSource(tn *ast.TableName) LogicalPlan {
 	var result LogicalPlan = ds
 	if b.ctx.Txn() != nil && !b.ctx.Txn().IsReadOnly() {
 		us := LogicalUnionScan{}.init(b.ctx)
-		us.SetSchema(ds.Schema().Clone())
 		us.SetChildren(result)
 		result = us
 	}
@@ -1756,7 +1750,6 @@ out:
 func (b *planBuilder) buildMaxOneRow(p LogicalPlan) LogicalPlan {
 	maxOneRow := LogicalMaxOneRow{}.init(b.ctx)
 	maxOneRow.SetChildren(p)
-	maxOneRow.SetSchema(p.Schema().Clone())
 	return maxOneRow
 }