--- a/plan/optimizer.go
+++ b/plan/optimizer.go
@@ -152,7 +152,6 @@ func dagPhysicalOptimize(logic LogicalPlan) (PhysicalPlan, error) {
 		return nil, errors.Trace(err)
 	}
 	p := t.plan()
-	rebuildSchema(p)
 	p.ResolveIndices()
 	return p, nil
 }