--- a/plan/physical_plans.go
+++ b/plan/physical_plans.go
@@ -365,15 +365,3 @@ type PhysicalTableDual struct {
 
 	RowCount int
 }
-
-func buildJoinSchema(joinType JoinType, join Plan) *expression.Schema {
-	switch joinType {
-	case SemiJoin, AntiSemiJoin:
-		return join.Children()[0].Schema().Clone()
-	case LeftOuterSemiJoin, AntiLeftOuterSemiJoin:
-		newSchema := join.Children()[0].Schema().Clone()
-		newSchema.Append(join.Schema().Columns[join.Schema().Len()-1])
-		return newSchema
-	}
-	return expression.MergeSchema(join.Children()[0].Schema(), join.Children()[1].Schema())
-}