refactor schema:
1. Only set schema for proj, apply, join, aggr, DataSource, Exists
2. Get Schema from its child.