@hanfei1991 2018-01-02
bench result

```
old bench:

BenchmarkOptimize/select_count(*)_from_t_group_by_e-4                              30000             48863 ns/op           36936 B/op        483 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_10_group_by_e-4                10000            108147 ns/op           76865 B/op        968 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_50-4                           10000            110577 ns/op           76969 B/op        968 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_c_>_'1'_group_by_b-4                10000            147542 ns/op           93881 B/op       1289 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_=_1_group_by_b-4                  10000            126735 ns/op           87593 B/op       1134 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_>_1_group_by_b-4                  10000            131862 ns/op           90745 B/op       1152 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_20-4                         10000            151946 ns/op           99345 B/op       1249 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_30-4                         10000            148137 ns/op           99345 B/op       1248 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_40-4                         10000            148687 ns/op           99346 B/op       1249 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_50-4                         10000            155522 ns/op           99346 B/op       1249 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_40-4                                30000             60077 ns/op           40640 B/op        477 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_50-4                                30000             59378 ns/op           40640 B/op        477 allocs/op
BenchmarkOptimize/select_*_from_t_where_1_and_t.b_<=_50-4                          30000             59484 ns/op           40944 B/op        484 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_100_order_by_t.a_limit_1-4          10000            208483 ns/op          153513 B/op       1740 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_1_order_by_t.a_limit_10-4           10000            204895 ns/op          152073 B/op       1704 allocs/op
BenchmarkOptimize/select_*_from_t_use_index(b)_where_b_=_1_order_by_a-4            30000             41726 ns/op           27560 B/op        350 allocs/op
BenchmarkOptimize/select_*_from_t_where_d_<_cast('1991-09-05'_as_datetime)-4               20000             65380 ns/op           40688 B/op        505 allocs/op
BenchmarkOptimize/select_*_from_t_where_ts_<_'1991-09-05'-4                                20000             68213 ns/op           40784 B/op        509 allocs/op

new bench

BenchmarkOptimize/select_count(*)_from_t_group_by_e-4                              30000             50367 ns/op           36936 B/op        483 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_10_group_by_e-4                10000            109644 ns/op           74817 B/op        953 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_50-4                           10000            110587 ns/op           74921 B/op        953 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_c_>_'1'_group_by_b-4                10000            143530 ns/op           91617 B/op       1273 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_=_1_group_by_b-4                  10000            124442 ns/op           85329 B/op       1118 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_>_1_group_by_b-4                  10000            131619 ns/op           88481 B/op       1136 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_20-4                         10000            144061 ns/op           97063 B/op       1232 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_30-4                         10000            146544 ns/op           97063 B/op       1232 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_40-4                         10000            144011 ns/op           97063 B/op       1232 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_50-4                         10000            148763 ns/op           97063 B/op       1232 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_40-4                                30000             56587 ns/op           38910 B/op        465 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_50-4                                30000             56218 ns/op           38910 B/op        465 allocs/op
BenchmarkOptimize/select_*_from_t_where_1_and_t.b_<=_50-4                          30000             58790 ns/op           39214 B/op        472 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_100_order_by_t.a_limit_1-4          10000            199191 ns/op          147161 B/op       1700 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_1_order_by_t.a_limit_10-4           10000            197413 ns/op          145681 B/op       1663 allocs/op
BenchmarkOptimize/select_*_from_t_use_index(b)_where_b_=_1_order_by_a-4            50000             39444 ns/op           24386 B/op        330 allocs/op
BenchmarkOptimize/select_*_from_t_where_d_<_cast('1991-09-05'_as_datetime)-4               20000             66211 ns/op           38968 B/op        494 allocs/op
BenchmarkOptimize/select_*_from_t_where_ts_<_'1991-09-05'-4                                20000             68685 ns/op           39064 B/op        498 allocs/op
```