@zimulala 2018-01-04
/run-all-tests