@winoros 2018-01-04
/run-integration-ddl-test